refactor: KEEP-356 consolidate payment protocols under lib/payments
mpp and x402 are payment protocols both consumed by lib/payments/
router.ts. Group them with the router instead of leaving them at
lib/ root.

Moves:
- lib/mpp/server.ts          -> lib/payments/mpp/server.ts
- lib/x402/execution-wait.ts -> lib/payments/x402/execution-wait.ts
- lib/x402/payment-gate.ts   -> lib/payments/x402/payment-gate.ts
- lib/x402/reconcile.ts      -> lib/payments/x402/reconcile.ts
- lib/x402/server.ts         -> lib/payments/x402/server.ts
- lib/x402/types.ts          -> lib/payments/x402/types.ts

lib/payments/router.ts unchanged at its existing location.
