+++ tests/unit/payment-router.test.ts
@@ -1,18 +1,18 @@
 import { describe, expect, it, vi } from "vitest";
 
 vi.mock("@x402/next", () => ({ withX402: vi.fn() }));
-vi.mock("@/lib/x402/payment-gate", () => ({
+vi.mock("@/lib/payments/x402/payment-gate", () => ({
   buildPaymentConfig: vi.fn(),
   extractPayerAddress: vi.fn(),
   findExistingPayment: vi.fn(),
   hashPaymentSignature: vi.fn(),
 }));
-vi.mock("@/lib/x402/server", () => ({ server: {} }));
-vi.mock("@/lib/x402/reconcile", () => ({
+vi.mock("@/lib/payments/x402/server", () => ({ server: {} }));
+vi.mock("@/lib/payments/x402/reconcile", () => ({
   isTimeoutError: vi.fn(),
   pollForPaymentConfirmation: vi.fn(),
 }));
-vi.mock("@/lib/mpp/server", () => ({
+vi.mock("@/lib/payments/mpp/server", () => ({
   extractMppPayerAddress: vi.fn(),
   hashMppCredential: vi.fn(),
   getMppServer: vi.fn(),
